+++ frontend/src/layouts/user/DefaultLayout.vue
@@ -8,8 +8,8 @@
 
 <script setup lang="ts">
 import { RouterView } from 'vue-router';
-import DefaultHeader from '@/layouts/DefaultHeader.vue';
-import DefaultFooter from '@/layouts/DefaultFooter.vue';
+import DefaultHeader from '@/layouts/user/DefaultHeader.vue';
+import DefaultFooter from '@/layouts/user/DefaultFooter.vue';
 </script>
 
 <style scoped>
