+++ frontend/src/router/index.ts
@@ -1,20 +1,22 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import HomeView from '../views/HomeView.vue';
-import DefaultLayout from '@/layouts/DefaultLayout.vue';
-import ProfileView from '@/views/ProfileView.vue';
-import PurchaseView from '@/views/PurchaseView.vue';
-import SellView from '@/views/SellView.vue';
-import AdminLogin from '@/views/AdminLogin.vue';
-import AddProduct from '@/views/AddProductView.vue';
-import EditProduct from '@/views/EditProductView.vue';
-import ProductView from '@/views/ProductView.vue';
-import NotFoundView from '@/views/NotFoundView.vue';
-import DefaultSide from '@/layouts/admin/DefaultSide.vue';
-import EditUserInfoView from '@/views/EditUserInfoView.vue';
-import EditProductByAdminView from '@/views/EditProductByAdminView.vue';
-import AddProductByAdminView from '@/views/AddProductByAdminView.vue';
-import OrderView from '@/views/OrderView.vue';
 
+import DefaultLayout from '@/layouts/user/DefaultLayout.vue';
+
+import HomeView from '@/views/user/HomeView.vue';
+import ProfileView from '@/views/user/profile/ProfileView.vue';
+import PurchaseView from '@/views/user/profile/PurchaseView.vue';
+import SellView from '@/views/user/profile/SellView.vue';
+import AddProduct from '@/views/user/AddProductView.vue';
+import EditProduct from '@/views/user/EditProductView.vue';
+import ProductView from '@/views/user/ProductView.vue';
+import NotFoundView from '@/views/user/NotFoundView.vue';
+import OrderView from '@/views/user/OrderView.vue';
+
+import AdminLogin from '@/views/admin/AdminLogin.vue';
+import DefaultSide from '@/layouts/admin/DefaultSide.vue';
+import EditUserInfoView from '@/views/admin/EditUserInfoView.vue';
+import EditProductByAdminView from '@/views/admin/EditProductByAdminView.vue';
+import AddProductByAdminView from '@/views/admin/AddProductByAdminView.vue';
 
 const router = createRouter({
   scrollBehavior() {
@@ -55,9 +57,9 @@ const router = createRouter({
       path: '/admin',
       name: 'In admin page',
       component: DefaultSide,
-      children:[
-        {path: 'product/edit/:id',name: 'admin edit product', component:EditProductByAdminView},
-        {path: 'product/add',name: 'admin add product', component:AddProductByAdminView}
+      children: [
+        { path: 'product/edit/:id', name: 'admin edit product', component: EditProductByAdminView },
+        { path: 'product/add', name: 'admin add product', component: AddProductByAdminView }
       ]
     }
   ]
