+++ frontend/src/views/admin/AddProductByAdminView.vue
