+++ frontend/src/views/admin/EditProductByAdminView.vue
