+++ frontend/src/views/user/AddProductView.vue
